Fix card error

+++ faefolkfantasium/routes.py
@@ -42,7 +42,7 @@ def edit_being(being_id):
 @app.route("/delete/<int:being_id>", methods=["POST"])
 def delete_being(being_id):
     being = Being.query.get_or_404(being_id)
-    db.session.delete(being)
-    db.session.commit()
-    return redirect(url_for("home"))
-
+    if request.method == "POST":
+        db.session.delete(being)
+        db.session.commit()
+        return redirect(url_for("home"))
